+++ modules/broker-auth/pom.xml
@@ -18,50 +18,27 @@
   ~
   -->
 
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <parent>
         <artifactId>message-broker</artifactId>
         <groupId>org.wso2.messaging</groupId>
-        <version>4.0.31-SNAPSHOT</version>
-        <relativePath>../..</relativePath>
+        <version>4.0.23-SNAPSHOT</version>
+        <relativePath>../../pom.xml</relativePath>
     </parent>
+
     <modelVersion>4.0.0</modelVersion>
     <packaging>bundle</packaging>
-    <artifactId>broker-amqp</artifactId>
-    <name>Broker - AMQP</name>
-
+    <artifactId>broker-auth</artifactId>
+    <name>Broker - Auth</name>
     <dependencies>
-        <dependency>
-            <groupId>org.wso2.messaging</groupId>
-            <artifactId>broker-core</artifactId>
-        </dependency>
         <dependency>
             <groupId>org.wso2.messaging</groupId>
             <artifactId>broker-common</artifactId>
         </dependency>
         <dependency>
-            <groupId>org.wso2.messaging</groupId>
-            <artifactId>broker-auth</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.wso2.messaging</groupId>
-            <artifactId>broker-coordination</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>io.netty</groupId>
-            <artifactId>netty-buffer</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>io.netty</groupId>
-            <artifactId>netty-transport</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>io.netty</groupId>
-            <artifactId>netty-codec</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>io.netty</groupId>
-            <artifactId>netty-handler</artifactId>
+            <groupId>org.wso2.carbon</groupId>
+            <artifactId>org.wso2.carbon.core</artifactId>
         </dependency>
         <dependency>
             <groupId>org.slf4j</groupId>
@@ -70,6 +47,10 @@
         <dependency>
             <groupId>org.slf4j</groupId>
             <artifactId>slf4j-log4j12</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>com.google.guava</groupId>
+            <artifactId>guava</artifactId>
         </dependency>
         <dependency>
             <groupId>com.google.code.findbugs</groupId>
@@ -85,6 +66,10 @@
             <groupId>org.mockito</groupId>
             <artifactId>mockito-core</artifactId>
             <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>org.wso2.carbon.config</groupId>
+            <artifactId>org.wso2.carbon.config</artifactId>
         </dependency>
     </dependencies>
 
@@ -116,24 +101,10 @@
                         <Private-Package>
                         </Private-Package>
                         <Import-Package>
-                            io.netty.bootstrap.*;version="${netty.import.package.range.version}",
-                            io.netty.buffer.*;version="${netty.import.package.range.version}",
-                            io.netty.channel.*;version="${netty.import.package.range.version}",
-                            io.netty.channel.nio.*;version="${netty.import.package.range.version}",
-                            io.netty.channel.socket.*;version="${netty.import.package.range.version}",
-                            io.netty.channel.socket.nio.*;version="${netty.import.package.range.version}",
-                            io.netty.handler.codec.*;version="${netty.import.package.range.version}",
-                            io.netty.handler.ssl.*;version="${netty.import.package.range.version}",
-                            io.netty.util.*;version="${netty.import.package.range.version}",
-                            io.netty.util.*.concurrent;version="${netty.import.package.range.version}",
-                            javax.net.ssl.*;version="0.0.0",
-                            org.wso2.broker.common.*;version="${broker.package.import.range.version}",
-                            org.wso2.broker.coordination.*; version="${broker.package.import.range.version}",
-                            org.wso2.broker.core.*;version="${broker.package.import.range.version}",
-                            org.slf4j.*;version="${slf4j.logging.package.import.version.range}",
+                            org.slf4j.*;version="${slf4j.logging.package.import.version.range}"
                         </Import-Package>
                         <Export-Package>
-                            org.wso2.broker.amqp.*; version="${project.version}"
+                            org.wso2.broker.auth.*;version=${project.version},
                         </Export-Package>
                     </instructions>
                 </configuration>
