+++ modules/broker-auth/src/main/java/org/wso2/broker/auth/authentication/sasl/plain/PlainSaslCallbackHandler.java
@@ -1,0 +1,78 @@
+/*
+ *   Copyright (c) 2018, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ *
+ *   WSO2 Inc. licenses this file to you under the Apache License,
+ *   Version 2.0 (the "License"); you may not use this file except
+ *   in compliance with the License.
+ *   You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *   Unless required by applicable law or agreed to in writing,
+ *   software distributed under the License is distributed on an
+ *   "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ *   KIND, either express or implied.  See the License for the
+ *   specific language governing permissions and limitations
+ *   under the License.
+ *
+ */
+package org.wso2.broker.auth.authentication.sasl.plain;
+
+import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
+
+import java.io.IOException;
+import javax.security.auth.callback.Callback;
+import javax.security.auth.callback.CallbackHandler;
+import javax.security.auth.callback.NameCallback;
+import javax.security.auth.callback.PasswordCallback;
+import javax.security.auth.callback.UnsupportedCallbackException;
+
+/**
+ * Class callback handler which is used to set user name and password given in incoming connection request.
+ * This will be used by {@link javax.security.auth.spi.LoginModule} to retrieve user
+ * authentication information to validate user
+ */
+public class PlainSaslCallbackHandler implements CallbackHandler {
+
+    private String username;
+    private char[] password;
+
+    @Override
+    public void handle(Callback[] callbacks) throws IOException, UnsupportedCallbackException {
+        if (callbacks != null) {
+            for (Callback callback : callbacks) {
+                if (callback instanceof NameCallback) {
+                    ((NameCallback) callback).setName(username);
+                } else if (callback instanceof PasswordCallback) {
+                    ((PasswordCallback) callback).setPassword(password);
+                } else {
+                    throw new UnsupportedCallbackException(callback);
+
+                }
+            }
+            clearCredentials();
+        }
+    }
+
+    /**
+     * Clear the credentials after handling data
+     */
+    private void clearCredentials() {
+        username = null;
+        if (password != null) {
+            for (int i = 0; i < password.length; i++) {
+                password[i] = ' ';
+            }
+            password = null;
+        }
+    }
+
+    public void setUsername(String username) {
+        this.username = username;
+    }
+
+    @SuppressFBWarnings("EI_EXPOSE_REP2")
+    public void setPassword(char... password) {
+        this.password = password;
+    }
+}
