+++ modules/broker-auth/src/main/java/org/wso2/broker/auth/user/UserStoreManager.java
@@ -1,0 +1,48 @@
+/*
+ *   Copyright (c) 2018, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ *
+ *   WSO2 Inc. licenses this file to you under the Apache License,
+ *   Version 2.0 (the "License"); you may not use this file except
+ *   in compliance with the License.
+ *   You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *   Unless required by applicable law or agreed to in writing,
+ *   software distributed under the License is distributed on an
+ *   "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ *   KIND, either express or implied.  See the License for the
+ *   specific language governing permissions and limitations
+ *   under the License.
+ *
+ */
+package org.wso2.broker.auth.user;
+
+import org.wso2.broker.auth.BrokerAuthException;
+
+import java.util.HashSet;
+
+/**
+ * Interface provides user store operations required for broker.
+ */
+public interface UserStoreManager {
+
+    /**
+     * Authenticate given user with credentials.
+     *
+     * @param userName    userName
+     * @param credentials credentials
+     * @return Authentication result
+     * @throws BrokerAuthException Exception throws when authentication failed.
+     */
+    public boolean authenticate(String userName, char... credentials) throws BrokerAuthException;
+
+    /**
+     * Retrieve the list of users for given username.
+     *
+     * @param userName user name
+     * @return List of roles
+     * @throws BrokerAuthException Exception throws when getting role list from user manager
+     */
+    public HashSet<String> getUserRoleList(String userName) throws BrokerAuthException;
+}
