+++ modules/broker-core/src/main/java/org/wso2/broker/core/Broker.java
@@ -21,20 +21,18 @@ package org.wso2.broker.core;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.wso2.broker.common.BrokerConfigProvider;
+import org.wso2.broker.auth.AuthManager;
 import org.wso2.broker.common.StartupContext;
 import org.wso2.broker.common.ValidationException;
 import org.wso2.broker.common.data.types.FieldTable;
 import org.wso2.broker.coordination.BasicHaListener;
 import org.wso2.broker.coordination.HaListener;
 import org.wso2.broker.coordination.HaStrategy;
-import org.wso2.broker.core.configuration.BrokerConfiguration;
 import org.wso2.broker.core.metrics.BrokerMetricManager;
 import org.wso2.broker.core.metrics.DefaultBrokerMetricManager;
 import org.wso2.broker.core.metrics.NullBrokerMetricManager;
 import org.wso2.broker.core.rest.api.ExchangesApi;
 import org.wso2.broker.core.rest.api.QueuesApi;
-import org.wso2.broker.core.security.authentication.AuthenticationManager;
 import org.wso2.broker.core.store.StoreFactory;
 import org.wso2.broker.rest.BrokerServiceRunner;
 import org.wso2.carbon.metrics.core.MetricService;
@@ -53,7 +51,7 @@ public final class Broker {
 
     private final MessagingEngine messagingEngine;
 
-    private final AuthenticationManager authenticationManager;
+    private final AuthManager authManager;
 
     /**
      * Used to manage metrics related to broker
@@ -81,10 +79,6 @@ public final class Broker {
         BrokerServiceRunner serviceRunner = startupContext.getService(BrokerServiceRunner.class);
         serviceRunner.deploy(new QueuesApi(this), new ExchangesApi(this));
         startupContext.registerService(Broker.class, this);
-        BrokerConfigProvider configProvider = startupContext.getService(BrokerConfigProvider.class);
-        BrokerConfiguration brokerConfiguration = configProvider
-                .getConfigurationObject(BrokerConfiguration.NAMESPACE, BrokerConfiguration.class);
-        this.authenticationManager = new AuthenticationManager(brokerConfiguration);
         haStrategy = startupContext.getService(HaStrategy.class);
         if (haStrategy == null) {
             brokerHelper = new BrokerHelper();
@@ -92,15 +86,16 @@ public final class Broker {
             LOGGER.info("Broker is in PASSIVE mode"); //starts up in passive mode
             brokerHelper = new HaEnabledBrokerHelper();
         }
+        this.authManager = startupContext.getService(AuthManager.class);
     }
 
     /**
-     * Provides {@link AuthenticationManager} for broker
+     * Provides {@link AuthManager} for broker
      *
      * @return Broker authentication manager
      */
-    public AuthenticationManager getAuthenticationManager() {
-        return authenticationManager;
+    public AuthManager getAuthManager() {
+        return authManager;
     }
 
     public void publish(Message message) throws BrokerException {
