+++ modules/broker-core/src/main/java/org/wso2/broker/core/configuration/BrokerConfiguration.java
@@ -19,9 +19,6 @@
 
 package org.wso2.broker.core.configuration;
 
-import java.util.HashMap;
-import java.util.Map;
-
 /**
  * Represents configuration for broker.
  */
@@ -43,22 +40,12 @@ public class BrokerConfiguration {
 
     private DataSourceConfiguration dataSource;
 
-    private AuthenticationConfiguration authenticator;
-    
     public DataSourceConfiguration getDataSource() {
         return dataSource;
     }
 
     public void setDataSource(DataSourceConfiguration dataSource) {
         this.dataSource = dataSource;
-    }
-
-    public AuthenticationConfiguration getAuthenticator() {
-        return authenticator;
-    }
-
-    public void setAuthenticator(AuthenticationConfiguration authenticator) {
-        this.authenticator = authenticator;
     }
 
     /**
@@ -113,36 +100,4 @@ public class BrokerConfiguration {
         }
 
     }
-
-    /**
-     * Represents a authentication configuration for broker ( e.g. database)
-     */
-    public static class AuthenticationConfiguration {
-        /**
-         * Jaas login module class name
-         */
-        private String loginModule;
-
-        /**
-         * Jaas login module options
-         */
-        private Map<String, String> options = new HashMap<>();
-
-        public String getLoginModule() {
-            return loginModule;
-        }
-
-        public void setLoginModule(String loginModule) {
-            this.loginModule = loginModule;
-        }
-
-        public Map<String, String> getOptions() {
-            return options;
-        }
-
-        public void setOptions(Map<String, String> options) {
-            this.options = options;
-        }
-    }
-
 }
