--- modules/broker-core/src/test/java/org/wso2/broker/core/security/authentication/sasl/plain/jaas/TestLoginModule.java
@@ -1,76 +1,0 @@
-/*
- *   Copyright (c) 2018, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
- *
- *   WSO2 Inc. licenses this file to you under the Apache License,
- *   Version 2.0 (the "License"); you may not use this file except
- *   in compliance with the License.
- *   You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- *   Unless required by applicable law or agreed to in writing,
- *   software distributed under the License is distributed on an
- *   "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- *   KIND, either express or implied.  See the License for the
- *   specific language governing permissions and limitations
- *   under the License.
- *
- */
-package org.wso2.broker.core.security.authentication.sasl.plain.jaas;
-
-import org.wso2.broker.core.security.authentication.exception.BrokerAuthenticationException;
-
-import java.io.IOException;
-import java.util.Arrays;
-import java.util.Map;
-import javax.security.auth.Subject;
-import javax.security.auth.callback.Callback;
-import javax.security.auth.callback.CallbackHandler;
-import javax.security.auth.callback.NameCallback;
-import javax.security.auth.callback.PasswordCallback;
-import javax.security.auth.callback.UnsupportedCallbackException;
-import javax.security.auth.login.LoginException;
-import javax.security.auth.spi.LoginModule;
-
-/**
- * Login module for unit tests
- */
-public class TestLoginModule implements LoginModule {
-    CallbackHandler callbackHandler;
-
-    @Override
-    public void initialize(Subject subject, CallbackHandler callbackHandler, Map<String, ?> sharedState,
-            Map<String, ?> options) {
-        this.callbackHandler = callbackHandler;
-    }
-
-    @Override
-    public boolean login() throws LoginException {
-        NameCallback userNameCallback = new NameCallback("userName");
-        PasswordCallback passwordCallback = new PasswordCallback("password", false);
-        Callback[] callbacks = { userNameCallback, passwordCallback };
-        try {
-            callbackHandler.handle(callbacks);
-        } catch (IOException | UnsupportedCallbackException e) {
-            throw new BrokerAuthenticationException("Error while handling callback ", e);
-        }
-        String userName = userNameCallback.getName();
-        char[] password = passwordCallback.getPassword();
-        return userName.equals("user") && Arrays.equals(password, new char[] { 'p', 'a', 's', 's' });
-    }
-
-    @Override
-    public boolean commit() throws LoginException {
-        return true;
-    }
-
-    @Override
-    public boolean abort() throws LoginException {
-        return false;
-    }
-
-    @Override
-    public boolean logout() throws LoginException {
-        return false;
-    }
-}
