+++ modules/coverage-report/pom.xml
@@ -92,5 +92,9 @@
             <groupId>org.wso2.messaging</groupId>
             <artifactId>broker-metrics</artifactId>
         </dependency>
+        <dependency>
+            <groupId>org.wso2.messaging</groupId>
+            <artifactId>broker-auth</artifactId>
+        </dependency>
     </dependencies>
 </project>
