+++ modules/launcher/src/main/assembly/assembly.xml
@@ -90,6 +90,7 @@
                 <include>org.wso2.messaging:broker-amqp:jar</include>
                 <include>org.wso2.messaging:broker-core:jar</include>
                 <include>org.wso2.messaging:broker-common:jar</include>
+                <include>org.wso2.messaging:broker-auth:jar</include>
                 <include>org.wso2.messaging:broker-coordination:jar</include>
                 <include>org.wso2.messaging:broker-launcher:jar</include>
             </includes>
