+++ modules/launcher/src/main/java/org/wso2/broker/Main.java
@@ -24,6 +24,10 @@ import com.zaxxer.hikari.HikariDataSource;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.wso2.broker.amqp.Server;
+import org.wso2.broker.auth.AuthManager;
+import org.wso2.broker.auth.BrokerAuthConfiguration;
+import org.wso2.broker.auth.user.UserStoreManager;
+import org.wso2.broker.auth.user.impl.UserStoreManagerImpl;
 import org.wso2.broker.common.BrokerConfigProvider;
 import org.wso2.broker.common.StartupContext;
 import org.wso2.broker.coordination.CoordinationException;
@@ -31,10 +35,6 @@ import org.wso2.broker.coordination.HaStrategy;
 import org.wso2.broker.coordination.HaStrategyFactory;
 import org.wso2.broker.core.Broker;
 import org.wso2.broker.core.configuration.BrokerConfiguration;
-import org.wso2.broker.core.security.authentication.user.User;
-import org.wso2.broker.core.security.authentication.user.UserStoreManager;
-import org.wso2.broker.core.security.authentication.user.UsersFile;
-import org.wso2.broker.core.security.authentication.util.BrokerSecurityConstants;
 import org.wso2.broker.metrics.BrokerMetricService;
 import org.wso2.broker.rest.BrokerRestServer;
 import org.wso2.carbon.config.ConfigProviderFactory;
@@ -43,7 +43,6 @@ import org.wso2.carbon.config.provider.ConfigProvider;
 
 import java.nio.file.Path;
 import java.nio.file.Paths;
-import java.util.List;
 import javax.sql.DataSource;
 
 /**
@@ -62,13 +61,17 @@ public class Main {
             StartupContext startupContext = new StartupContext();
 
             initConfigProvider(startupContext);
-            loadUsers();
             BrokerConfigProvider service = startupContext.getService(BrokerConfigProvider.class);
             BrokerConfiguration brokerConfiguration =
                     service.getConfigurationObject(BrokerConfiguration.NAMESPACE, BrokerConfiguration.class);
             DataSource dataSource = getDataSource(brokerConfiguration.getDataSource());
             startupContext.registerService(DataSource.class, dataSource);
-
+            // Initializing broker auth module
+            UserStoreManager userStoreManager = new UserStoreManagerImpl();
+            BrokerAuthConfiguration securityConfiguration = service
+                    .getConfigurationObject(BrokerAuthConfiguration.NAMESPACE, BrokerAuthConfiguration.class);
+            startupContext.registerService(AuthManager.class, new AuthManager(securityConfiguration, dataSource,
+                                                                              userStoreManager));
             HaStrategy haStrategy;
             //Initializing an HaStrategy implementation only if HA is enabled
             try {
@@ -139,29 +142,6 @@ public class Main {
         ConfigProvider configProvider = ConfigProviderFactory.getConfigProvider(brokerYamlFile, null);
         startupContext.registerService(BrokerConfigProvider.class,
                                        (BrokerConfigProvider) configProvider::getConfigurationObject);
-    }
-
-    /**
-     * Loads the users from users.yaml during broker startup
-     */
-    private static void loadUsers() throws ConfigurationException {
-        Path usersYamlFile;
-        String usersFilePath = System.getProperty(BrokerSecurityConstants.SYSTEM_PARAM_USERS_CONFIG);
-        if (usersFilePath == null || usersFilePath.trim().isEmpty()) {
-            // use current path.
-            usersYamlFile = Paths.get("", BrokerSecurityConstants.USERS_FILE_NAME).toAbsolutePath();
-        } else {
-            usersYamlFile = Paths.get(usersFilePath).toAbsolutePath();
-        }
-        ConfigProvider configProvider = ConfigProviderFactory.getConfigProvider(usersYamlFile, null);
-        UsersFile usersFile = configProvider
-                .getConfigurationObject(BrokerSecurityConstants.USERS_CONFIG_NAMESPACE, UsersFile.class);
-        if (usersFile != null) {
-            List<User> users = usersFile.getUsers();
-            for (User user : users) {
-                UserStoreManager.addUser(user);
-            }
-        }
     }
 
     /**
