+++ pom.xml
@@ -27,7 +27,7 @@
         <version>5</version>
     </parent>
 
-    <groupId>org.wso2.messaging</groupId> 
+    <groupId>org.wso2.messaging</groupId>
     <artifactId>message-broker</artifactId>
     <packaging>pom</packaging>
     <version>4.0.31-SNAPSHOT</version>
@@ -43,6 +43,7 @@
 
     <modules>
         <module>modules/broker-common</module>
+        <module>modules/broker-auth</module>
         <module>modules/broker-metrics</module>
         <module>modules/broker-rest-runner</module>
         <module>modules/broker-core</module>
@@ -89,6 +90,11 @@
             <dependency>
                 <groupId>org.wso2.messaging</groupId>
                 <artifactId>broker-metrics</artifactId>
+                <version>${project.version}</version>
+            </dependency>
+            <dependency>
+                <groupId>org.wso2.messaging</groupId>
+                <artifactId>broker-auth</artifactId>
                 <version>${project.version}</version>
             </dependency>
             <dependency>
@@ -176,6 +182,11 @@
                 <groupId>org.apache.geronimo.specs.wso2</groupId>
                 <artifactId>geronimo-jms_1.1_spec</artifactId>
                 <version>${geronimo-jms_1.1_spec.version}</version>
+            </dependency>
+            <dependency>
+                <groupId>org.wso2.carbon</groupId>
+                <artifactId>org.wso2.carbon.core</artifactId>
+                <version>${carbon.kernel.version}</version>
             </dependency>
             <dependency>
                 <groupId>org.wso2.carbon.config</groupId>
@@ -315,6 +326,7 @@
 
     <properties>
         <broker.package.import.range.version>[4.0.0,5.0.0)</broker.package.import.range.version>
+        <carbon.kernel.version>5.2.7</carbon.kernel.version>
         <netty.version>4.1.17.Final</netty.version>
         <netty.import.package.range.version>(4,5]</netty.import.package.range.version>
         <slf4j.version>1.7.25</slf4j.version>
